Added tests for dropout in test_functional.py
Removed tests for dropout from test_precision.py
